multi_agent_finance_assistant/requirements.txt
streamlit
fastapi
uvicorn
yfinance
beautifulsoup4
faiss-cpu
openai
pyttsx3
whisper
beautifulsoup4
requests

